Basic blog rss feeds, multiple markup types are now supported for blog posts, pygments for code highliting, fixed some slug-related bugs, teasers are no longer stored in the db

+++ mymodernlife/apps/blog/views.py
@@ -1,6 +1,6 @@
 import datetime
 
-from django.shortcuts import render_to_response, redirect
+from django.shortcuts import render_to_response, redirect, get_object_or_404
 from django.template import RequestContext
 from django.contrib.auth.decorators import login_required
 from django.http import Http404
@@ -15,11 +15,8 @@ TODO:
     - Even better: have contributors be a group instead of users list?
     - Make this even more portable: stick templates as a parameter
     - Pass blog ownership to a new user
-    - Feeds
     - Linkbacks/Pingbacks/Trackbacks?
-    - Multiple input formats (wsywig, markdown, rest, plaintext, etc)
     - Import/export tools
-    - pygments for code highlighting
 """
 
 def index(request):
@@ -56,32 +53,25 @@ def create_blog(request):
     return render_to_response('blog/create_blog.html', context, RequestContext(request))
 
 def view_blog(request, slug):
-    blog = Blog.objects.filter(slug=slug)
+    blog = get_object_or_404(Blog, slug=slug)
     posts = Post.objects.filter(blog=blog)
 
-    if not blog:
-        raise Http404
-
     context = {
-        'blog': blog[0],
+        'blog': blog,
         'posts': posts,
     }
     return render_to_response('blog/view_blog.html', context, RequestContext(request))
 
 @login_required
 def delete_blog(request, slug):
-    blog = Blog.objects.filter(slug=slug)
+    if request.method == 'POST':
+        blog = get_object_or_404(Blog, slug=slug)
 
-    if not blog:
-        raise Http404
+        # Users are only allowed to delete a blog that they own
+        if request.user.username != blog.owner.username:
+            return redirect('view_blog', blog.slug)
 
-    blog = blog[0]
-
-    # Users are only allowed to delete a blog that they own
-    if request.user.username != blog.owner.username:
-        return redirect('view_blog', blog.slug)
-
-    blog.delete()
+        blog.delete()
 
     return redirect(index)
 
@@ -159,16 +149,17 @@ def edit_post(request, year, month, slug):
 
 @login_required
 def delete_post(request, year, month, slug):
-    post = Post.objects.filter(created__year=year, created__month=month, slug=slug)
-    post = post[0]
+    if request.method == 'POST':
+        post = Post.objects.filter(created__year=year, created__month=month, slug=slug)
+        post = post[0]
 
-    blog = Blog.objects.get(id=post.blog.id)
+        blog = Blog.objects.get(id=post.blog.id)
 
-    # Users are only allowed to delete their own blog posts
-    if request.user.username != post.author.username:
-        return redirect('view_post', post.created.year, post.get_formatted_month(), post.slug)
+        # Users are only allowed to delete their own blog posts
+        if request.user.username != post.author.username:
+            return redirect('view_post', post.created.year, post.get_formatted_month(), post.slug)
 
-    post.delete()
+        post.delete()
 
     return redirect('view_blog', blog.slug)
